+++ src/components/user/Home/UserHome.scss
@@ -1,5 +1,5 @@
-@use "../../styles/function" as *;
-@use "../../styles/variables" as *;
+@use "../../../styles/function" as *;
+@use "../../../styles/variables" as *;
 
 .user-home {
   width: 100%;
