+++ src/pages/user/UserHomePage.jsx
@@ -1,5 +1,5 @@
 import { Footer } from "@/components/common/Footer";
-import { UserHome } from "@/components/user/UserHome";
+import { UserHome } from "@/components/user/Home/UserHome";
 
 export const UserHomePage = () => {
   return (
